refactor: unite all substituiontests in the correct place
